requirements.txt
django
pillow
psycopg2-binary
requests
dj-database-url
uvicorn
django_crondj-database-url==3.0.1
